+++ src/WindowsDeviceManagerViewer/Utilities/DatabaseReader.cs
@@ -20,6 +20,7 @@ namespace WindowsDeviceManagerViewer.Utilities
 
             if (File.Exists(databasefile))
             {
+                string databaseVersion = GetDatabaseUserVersion(databasefile);
                 using SQLiteConnection connection = new($"Data Source = {databasefile}");
                 connection.Open();
                 using (SQLiteCommand command = connection.CreateCommand())
@@ -28,15 +29,51 @@ namespace WindowsDeviceManagerViewer.Utilities
                     using var executeReader = command.ExecuteReader();
                     while (executeReader.Read())
                     {
-                        WindowsDeviceInfo readRecord = new()
+                        WindowsDeviceInfo readRecord;
+                        switch (databaseVersion)
                         {
-                            HostName = executeReader["HostName"].ToString(),
-                            UserName = executeReader["UserName"].ToString(),
-                            OSName = executeReader["OSName"].ToString(),
-                            OSBuildNumber = executeReader["OSBuildNumber"].ToString(),
-                            OSVersion = executeReader["OSVersion"].ToString(),
-                            LastUpdate = executeReader["LastUpdate"].ToString()
-                        };
+                            case "0":       // user_versionが0のとき
+                                readRecord = new()
+                                {
+                                    HostName = executeReader["HostName"].ToString(),
+                                    UserName = executeReader["UserName"].ToString(),
+                                    OSName = executeReader["OSName"].ToString(),
+                                    OSBuildNumber = executeReader["OSBuildNumber"].ToString(),
+                                    OSVersion = executeReader["OSVersion"].ToString(),
+                                    ComputerManufacturer = string.Empty,
+                                    ComputerModel = string.Empty,
+                                    LastUpdate = executeReader["LastUpdate"].ToString()
+                                };
+                                break;
+
+                            case "1":       // user_versionが1のとき
+                                readRecord = new()
+                                {
+                                    HostName = executeReader["HostName"].ToString(),
+                                    UserName = executeReader["UserName"].ToString(),
+                                    OSName = executeReader["OSName"].ToString(),
+                                    OSBuildNumber = executeReader["OSBuildNumber"].ToString(),
+                                    OSVersion = executeReader["OSVersion"].ToString(),
+                                    ComputerManufacturer = executeReader["ComputerManufacturer"].ToString(),
+                                    ComputerModel = executeReader["ComputerModel"].ToString(),
+                                    LastUpdate = executeReader["LastUpdate"].ToString()
+                                };
+                                break;
+
+                            default:        // user_versionが想定外のとき
+                                readRecord = new()
+                                {
+                                    HostName = string.Empty,
+                                    UserName = string.Empty,
+                                    OSName = string.Empty,
+                                    OSBuildNumber = string.Empty,
+                                    OSVersion = string.Empty,
+                                    ComputerManufacturer = string.Empty,
+                                    ComputerModel = string.Empty,
+                                    LastUpdate = string.Empty
+                                };
+                                break;
+                        }
                         readRecords.Add(readRecord);
                     }
                 }
@@ -44,6 +81,27 @@ namespace WindowsDeviceManagerViewer.Utilities
             }
 
             return readRecords;
+        }
+
+        /// <summary>
+        /// データベースuser_version取得処理
+        /// </summary>
+        /// <param name="databasefile">データベースファイル名</param>
+        /// <returns>データベースのuser_version</returns>
+        private static string GetDatabaseUserVersion(string databasefile)
+        {
+            string version = string.Empty;
+
+            using SQLiteConnection connection = new($"Data Source = {databasefile}");
+            connection.Open();
+            using (SQLiteCommand command = connection.CreateCommand())
+            {
+                command.CommandText = "PRAGMA user_version";
+                version = command.ExecuteScalar().ToString();
+            }
+            connection.Close();
+
+            return version;
         }
     }
 }